feat: split manage network model
