+++ src/renderer/features/network/NetworkSelector/index.ts
@@ -1,2 +1,3 @@
 export { NetworkSelector } from './ui/NetworkSelector';
 export { networkSelectorUtils } from './lib/network-selector-utils';
+export { networkSelectorModel } from './model/network-selector-model';
